Improve NSCA reliability when having multiple receivers
When DNS entry for receiver has multiple IPs the behavior of selecting a
random one, failing to connect and wait for 10 seconds is inefficient.
It results in many checks being stale when one of the IP does not
respond.
This commit tries all the IPs before backing off and gives the priority
to the last known good IP in order to make failover smoother.