Improve network plugin execution time
When having the 4 networks checks (data up/down and packets up/down) on
an interface, Sauna used to execute them sequentially. It now executes
it once and cache the result for the other checks of the same interface.

It means that instead of taking 4 seconds to execute, they now only take
1 second.